@@ -5,11 +5,11 @@ homepage: https://gitlab.com/tobiaswkjeldsen/dartnissanconnect
 author: Tobias Westergaard Kjeldsen <tobias@wkjeldsen.dk>
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  intl: '^0.15.7'
-  http: '^0.12.0'
+  intl: '^0.17.0'
+  http: '^0.13.1'
 
 dev_dependencies:
   test: '^1.3.0'
